+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - fixed GUI.msgbox_with_checkbox
     - fixed SpectrometerApplicationState.check_refs
     - moved plugin persistence from "advanced_options" to "plugins"
+    - SRM works properly in Presets
 - 2024-03-22 4.0.55
     - added msgbox_with_lineedit
     - hopefully didn't break dialog response processing